refactor(emitter): split printer setup from core
